requirements.txt
aldjemy==0.6.0
appdirs==1.4.0
dj-database-url==0.4.2
Django==1.10.5
django-appconf==1.0.2
django-bootstrap3==8.2.1
django-crispy-forms==1.6.1
django-mathfilters==0.4.0
django-postman==3.5.1
django-render-block==0.5
django-tables2==1.4.2
django-templated-email==2.1
django-webpack-loader==0.2.4
djangorestframework==3.5.4
gunicorn==19.7.0
lockfile==0.12.2
optional-django==0.3.0
packaging==16.8
psycopg2==2.6.2
pyparsing==2.1.10
python-decouple==3.0
requests==2.13.0
six==1.10.0
SQLAlchemy==1.1.5
webpack==6.0.0
whitenoise==3.3.0
